// enums.ts
 export enum Tab {
   Explore = 'explore',
   Marketplace = 'marketplace',
   Events = 'events',
   Profile = 'profile',
   Play = 'play',
 }
 
 export enum View {
   Explore = 'explore',
   CraftDetail = 'craftDetail',
   AiStudio = 'aiStudio',
   EventDetail = 'eventDetail',
   ProductDetail = 'productDetail',
   Chatroom = 'chatroom',
   TextLab = 'textLab',
 }
 
 export enum ArtisanView {
     List = 'list',
     Chatroom = 'chatroom',
 }
 
 export enum ArtisanTab {
     Dashboard = 'dashboard',
     Products = 'products',
     Orders = 'orders',
     Messages = 'messages',
     Settings = 'settings'
 }